[REFACTOR] Certik Audit Recommendation - Exhibit 27
Description:
Throughout the codebase the variables “snapshotIndex” as well as “lastSnapshotIndex” are
declared as being of type “int” when representing an array index. Additionally, the function
“_findDividendsSnapshot” returns an “int” representation of the index.

Recommendations:
An “overflow” of the index variable can occur in the “hypothetical” scenario that the snapshots
are generated in perpetuity, as a “uint” is able to represent numbers up to 2^256 - 1 whereas an
“int” is able to represent numbers up to 2^255 - 1. Should such a high number of snapshots be
achieved, the contract would lock whereas if “uint”s are used the contract would simply “reset”
at 2^256 - 1 entries.

Regardless of the hypothetical scenario, the representation of a non-negative index as a signed
integer goes against the best coding practices and as such we suggest the issue to be
remediated by converting the return value of “_findDividendsSnapshot” to a “uint” or changing all
variables contained in the function to “uint”s.