[CHANGED] Certik Audit - Exhibit 6 (round 3)
[INFORMATIONAL] Description:
On each “setRewardsForPeriods” call, a multiplication is performed with the “immutable” variable “periodLengthInCycles” and the result of all operations is added to the “totalPrizePool”.

Recommendations:
As the multiplication is done with an “immutable” variable that remains the same across invocations, it is possible to omit it and instead conduct a single multiplication on the result of “totalPrizePool” where it is used on L193. This will reduce the total gas cost necessary for initializing the contract.