[CHANGED] Certik Audit - Exhibit 8 (round 3)
[MINOR] Description:
The “setRewardsForPeriods” function can be called at any time on the contract, allowing the owner to overwrite previously set rewards at will.

Recommendations:
The same protection method as Exhibit 7 should be applied, as this function enables the “owner” to in essence “lure” numerous NFT stakes on the platform and subsequently nullify their rewards.