[CHANGED] Certik Audit - Exhibit 9 (round 3)
[INFORMATIONAL] Description:
The “_updateHistory” function returns a “Snapshot” struct copied via “memory” whilst only a single member of the struct is actually utilized.

Recommendations:
As the function is “internal” and does not exist anywhere else in the codebase, it would be beneficial gas-wise to instead return the “stake” member’s value of a “Snapshot” directly rather than copying the whole struct in memory unless these methods are planned to be utilized by derivative contracts.